% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs.R
\name{cvs_qc}
\alias{cvs_qc}
\title{CVS QC Interface}
\usage{
cvs_qc(main_path)
}
\arguments{
\item{main_path}{Provide the path where you saved the results from the first stage.}
}
\value{
A web link to start the interactive QC session.
}
\description{
Start the R Shiny with a CVS QC interface.
}
