% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_prep.R
\name{get_labeled_mask}
\alias{get_labeled_mask}
\title{Label Lesions}
\usage{
get_labeled_mask(mask)
}
\arguments{
\item{mask}{The binary lesion mask.}
}
\value{
A labeled lesion image.
}
\description{
Generate lesion labels for a binary lesion mask.
}
