% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JLF.R
\name{jlf_qc}
\alias{jlf_qc}
\title{JLF ROI QC Interface}
\usage{
jlf_qc(main_path)
}
\arguments{
\item{main_path}{Provide the path where you saved the results from the first stage.}
}
\value{
A web link to start the interactive QC session.
}
\description{
Start the R Shiny with a JLF ROI QC interface.
}
