% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesion.R
\name{lesion_qc}
\alias{lesion_qc}
\title{WM Lesion QC Interface}
\usage{
lesion_qc(main_path)
}
\arguments{
\item{main_path}{Provide the path where you saved the results from the first stage.}
}
\value{
A web link to start the interactive QC session.
}
\description{
Start the R Shiny with a WM lesion QC interface.
}
