% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_QC.R
\name{post_qc}
\alias{post_qc}
\title{Post-QC Interface}
\usage{
post_qc(main_path)
}
\arguments{
\item{main_path}{Provide the path where you saved the QC results.}
}
\value{
A web link to start the interactive QC session.
}
\description{
Start the R Shiny application with the post-QC interface.
}
