% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRL.R
\name{prl_qc}
\alias{prl_qc}
\title{PRL Lesion QC Interface}
\usage{
prl_qc(main_path)
}
\arguments{
\item{main_path}{Provide the path where you saved the results from the first stage.}
}
\value{
A web link to start the interactive QC session.
}
\description{
Start the R Shiny with a PRL lesion QC interface.
}
