% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_tool.R
\name{qc_shiny}
\alias{qc_shiny}
\title{Brain QC Shiny}
\usage{
qc_shiny(qc_type = "lesion", main_path)
}
\arguments{
\item{qc_type}{Specify the type of QC procedure: lesion, freesurfer, JLF, PRL, cvs.}

\item{main_path}{Provide the path where you saved the results from the first stage.}
}
\value{
\code{qc_type} returns a web link to start the interactive QC session.
}
\description{
Start the Interactive QC Session through Rshiny.
}
