% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_prep.R
\name{read_mgz}
\alias{read_mgz}
\title{Transform Image Types}
\usage{
read_mgz(file)
}
\arguments{
\item{file}{The mgz file path.}
}
\value{
An image in nifti format.
}
\description{
Transform mgz image type to nifti image type.
}
