% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_prep.R
\name{stage1_qc}
\alias{stage1_qc}
\title{Brain QC Preparation}
\usage{
stage1_qc(
  main_path,
  img_name,
  seg_name,
  subject,
  cores = 1,
  qc_type = "lesion",
  default_seg = NULL
)
}
\arguments{
\item{main_path}{Provide the path to neuroimaging data in BIDS format.}

\item{img_name}{The brain image name.}

\item{seg_name}{The segmentation image name.}

\item{subject}{Subject ID.}

\item{cores}{Number of cores used for parallel computing.}

\item{qc_type}{Specify the type of QC procedure: lesion, freesurfer, JLF, PRL, cvs.}

\item{default_seg}{Select a default ROI to be evaluated first (when choosing freesurfer or JLF as the \code{qc_type}).}
}
\value{
\code{stage1_qc} returns a list containing the images to be evaluated as well as the evaluation form.
}
\description{
Prepare brain images and segmentation images for evaluation.
}
